+++ src/pages/home/subPages/components/memo/MemoSearchTextArea/MemoSearchTextArea.tsx
@@ -1,25 +1,7 @@
 import { ChangeEvent, useState } from 'react';
 import { TextareaAutosize } from '@mui/material';
 import { SearchIcon } from 'assets/icons';
-
-import { KeyboardEvent } from 'react';
-
-// Enter keyboard 동작 확인 시, submit 동작 수행
-// form 형식은 textarea 못 써서, keyboard 입력으로 직접 수행
-const usePressEnterFetch = ({ handleSubmit }: { handleSubmit: () => void }) => {
-  const handlePressEnterFetch = (e: KeyboardEvent<HTMLTextAreaElement>) => {
-    if (e.nativeEvent.isComposing) {
-      return;
-    }
-
-    if (e.key === 'Enter' && !e.shiftKey) {
-      e.preventDefault();
-      handleSubmit();
-    }
-  };
-
-  return { handlePressEnterFetch };
-};
+import { usePressEnterFetch } from 'pages/home/subPages/hooks';
 
 interface MemoSearchTextAreaProps {
   value: string;
