+++ src/pages/home/subPages/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as usePressEnterFetch } from './usePressEnterFetch';
